refactor to use useAutocomplete hook, remove async stuff for now, will address in another item
